pyproject.toml
[project]
name = "knowledge-base-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.119.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.37.0",
]
